backend/requirements.txt
fastapi[all]
uvicorn[standard]
supabase
google-genai>=0.2.0
python-dotenv
python-multipart
pillow
pyjwt

